Simplifying max-min-exact logic, exporting decorator
